[refactor] folder structure for context utils
